backend-test/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.8.2
PyMuPDF==1.24.9
pillow==10.4.0
pytesseract==0.3.13
langgraph==0.2.14
langchain==0.2.14
langchain-google-genai==1.0.10
python-dotenv==1.0.1
